video-editor/requirements.txt
absl-py==2.1.0
bidict==0.23.1
boto3==1.35.5
botocore==1.35.5
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
decorator==4.4.2
et-xmlfile==1.1.0
filelock==3.15.4
filetype==1.2.0
fire==0.6.0
fonttools==4.53.1
fsspec==2024.6.1
git-filter-repo==2.45.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.1
google-cloud-videointelligence==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
grpcio==1.67.1
grpcio-status==1.67.1
huggingface-hub==0.24.6
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
jmespath==1.0.1
keyboard==0.13.5
kiwisolver==1.4.5
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
moviepy==1.0.3
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
proglog==0.1.10
proto-plus==1.24.0
protobuf==5.27.3
psutil==6.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybboxes==0.1.6
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.42
rsa==4.9
s3transfer==0.10.2
sahi==0.11.18
scipy==1.14.1
seaborn==0.13.2
setuptools==73.0.1
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.82
ultralytics-thop==2.0.5
urllib3==2.2.2
Werkzeug==3.0.4
yolov5==7.0.13